+++ src/main/java/calculator/common/io/IOManager.java
@@ -1,7 +1,7 @@
-package calculator.model.io;
+package calculator.common.io;
 
 public interface IOManager {
     default void printMessage(String message) {
         System.out.println(message);
-    };
+    }
 }
